feat: create messaging and editor packages
