+++ packages/editor/src/index.js
@@ -1,0 +1,2 @@
+import WrappedEditorView from './view';
+export default WrappedEditorView;
