+++ packages/messaging/package.json
@@ -1,0 +1,32 @@
+{
+  "name": "@nteract/messaging",
+  "version": "1.0.0",
+  "description": "Messaging mechanics for nteract apps (jupyter spec)",
+  "main": "lib/index.js",
+  "scripts": {
+    "prepublish": "npm run build",
+    "build": "npm run build:clean && npm run build:lib && npm run build:flow",
+    "build:clean": "rimraf lib",
+    "build:flow": "flow-copy-source -v -i '**/__tests__/**' src lib",
+    "build:lib": "babel -d lib src --ignore '**/__tests__/**'",
+    "test": "jest"
+  },
+  "dependencies": {
+    "rxjs": "^5.0.2"
+  },
+  "devDependencies": {
+    "babel-cli": "^6.23.0",
+    "flow-copy-source": "^1.1.0",
+    "jest": "^18.1.0",
+    "@nteract/dev-scripts": "^1.0.0",
+    "@nteract/mockument": "^1.0.0"
+  },
+  "keywords": [
+    "nteract",
+    "messaging",
+    "jmp",
+    "jupyter"
+  ],
+  "author": "Kyle Kelley <rgbkrk@gmail.com>",
+  "license": "BSD-3-Clause"
+}
